Move kernel function constructors to their respective algorithm-class folder
