+++ blackjax/__init__.py
@@ -1,30 +1,26 @@
 from blackjax._version import __version__
 
+from .adaptation.meads_adaptation import meads_adaptation
+from .adaptation.pathfinder_adaptation import pathfinder_adaptation
+from .adaptation.window_adaptation import window_adaptation
 from .diagnostics import effective_sample_size as ess
 from .diagnostics import potential_scale_reduction as rhat
-from .kernels import (
-    adaptive_tempered_smc,
-    additive_step_random_walk,
-    csgld,
-    elliptical_slice,
-    ghmc,
-    hmc,
-    irmh,
-    mala,
-    meads_adaptation,
-    meanfield_vi,
-    mgrad_gaussian,
-    nuts,
-    orbital_hmc,
-    pathfinder,
-    pathfinder_adaptation,
-    rmh,
-    sghmc,
-    sgld,
-    tempered_smc,
-    window_adaptation,
-)
+from .mcmc.elliptical_slice import elliptical_slice
+from .mcmc.ghmc import ghmc
+from .mcmc.hmc import hmc
+from .mcmc.mala import mala
+from .mcmc.marginal_latent_gaussian import mgrad_gaussian
+from .mcmc.nuts import nuts
+from .mcmc.periodic_orbital import orbital_hmc
+from .mcmc.random_walk import additive_step_random_walk, irmh, rmh
 from .optimizers import dual_averaging, lbfgs
+from .sgmcmc.csgld import csgld
+from .sgmcmc.sghmc import sghmc
+from .sgmcmc.sgld import sgld
+from .smc.adaptive_tempered import adaptive_tempered_smc
+from .smc.tempered import tempered_smc
+from .vi.meanfield_vi import meanfield_vi
+from .vi.pathfinder import pathfinder
 
 __all__ = [
     "__version__",
