+++ blackjax/adaptation/pathfinder_adaptation.py
@@ -12,24 +12,120 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Implementation of the Pathinder warmup for the HMC family of sampling algorithms."""
-from typing import NamedTuple, Tuple
+from typing import Callable, NamedTuple, Tuple, Union
 
+import jax
 import jax.numpy as jnp
 
+import blackjax.mcmc as mcmc
+import blackjax.vi as vi
+from blackjax.adaptation.base import AdaptationInfo, AdaptationResults
 from blackjax.adaptation.step_size import (
     DualAveragingAdaptationState,
     dual_averaging_adaptation,
 )
+from blackjax.base import AdaptationAlgorithm
 from blackjax.optimizers.lbfgs import lbfgs_inverse_hessian_formula_1
-from blackjax.types import Array, PyTree
+from blackjax.types import Array, PRNGKey, PyTree
 
-__all__ = ["PathfinderAdaptationState", "base"]
+__all__ = ["PathfinderAdaptationState", "base", "pathfinder_adaptation"]
 
 
 class PathfinderAdaptationState(NamedTuple):
     ss_state: DualAveragingAdaptationState
     step_size: float
     inverse_mass_matrix: Array
+
+
+def pathfinder_adaptation(
+    algorithm: Union[mcmc.hmc.hmc, mcmc.nuts.nuts],
+    logdensity_fn: Callable,
+    initial_step_size: float = 1.0,
+    target_acceptance_rate: float = 0.80,
+    **extra_parameters,
+) -> AdaptationAlgorithm:
+    """Adapt the value of the inverse mass matrix and step size parameters of
+    algorithms in the HMC fmaily.
+
+    Parameters
+    ----------
+    algorithm
+        The algorithm whose parameters are being tuned.
+    logdensity_fn
+        The log density probability density function from which we wish to sample.
+    initial_step_size
+        The initial step size used in the algorithm.
+    target_acceptance_rate
+        The acceptance rate that we target during step size adaptation.
+    **extra_parameters
+        The extra parameters to pass to the algorithm, e.g. the number of
+        integration steps for HMC.
+
+    Returns
+    -------
+    A function that returns the last chain state and a sampling kernel with the
+    tuned parameter values from an initial state.
+
+    """
+
+    mcmc_kernel = algorithm.build_kernel()
+
+    adapt_init, adapt_update, adapt_final = base(
+        target_acceptance_rate,
+    )
+
+    def one_step(carry, rng_key):
+        state, adaptation_state = carry
+        new_state, info = mcmc_kernel(
+            rng_key,
+            state,
+            logdensity_fn,
+            adaptation_state.step_size,
+            adaptation_state.inverse_mass_matrix,
+            **extra_parameters,
+        )
+        new_adaptation_state = adapt_update(
+            adaptation_state, new_state.position, info.acceptance_rate
+        )
+        return (
+            (new_state, new_adaptation_state),
+            AdaptationInfo(new_state, info, new_adaptation_state),
+        )
+
+    def run(rng_key: PRNGKey, position: PyTree, num_steps: int = 400):
+        init_key, sample_key, rng_key = jax.random.split(rng_key, 3)
+
+        pathfinder_state, _ = vi.pathfinder.approximate(
+            init_key, logdensity_fn, position
+        )
+        init_warmup_state = adapt_init(
+            pathfinder_state.alpha,
+            pathfinder_state.beta,
+            pathfinder_state.gamma,
+            initial_step_size,
+        )
+
+        init_position, _ = vi.pathfinder.sample(sample_key, pathfinder_state)
+        init_state = algorithm.init(init_position, logdensity_fn)
+
+        keys = jax.random.split(rng_key, num_steps)
+        last_state, info = jax.lax.scan(
+            one_step,
+            (init_state, init_warmup_state),
+            keys,
+        )
+        last_chain_state, last_warmup_state = last_state
+
+        step_size, inverse_mass_matrix = adapt_final(last_warmup_state)
+        parameters = {
+            "step_size": step_size,
+            "inverse_mass_matrix": inverse_mass_matrix,
+            **extra_parameters,
+        }
+
+        return AdaptationResults(last_chain_state, parameters), info
+
+    return AdaptationAlgorithm(run)
 
 
 def base(
