+++ blackjax/mcmc/elliptical_slice.py
@@ -17,10 +17,17 @@ from typing import Callable, NamedTuple, Tuple
 import jax
 import jax.numpy as jnp
 
+from blackjax.base import MCMCSamplingAlgorithm
 from blackjax.types import Array, PRNGKey, PyTree
 from blackjax.util import generate_gaussian_noise
 
-__all__ = ["EllipSliceState", "EllipSliceInfo", "init", "build_kernel"]
+__all__ = [
+    "EllipSliceState",
+    "EllipSliceInfo",
+    "init",
+    "build_kernel",
+    "elliptical_slice",
+]
 
 
 class EllipSliceState(NamedTuple):
@@ -110,6 +117,64 @@ def build_kernel(cov_matrix: Array, mean: Array):
         return proposal_generator(rng_key, state)
 
     return kernel
+
+
+class elliptical_slice:
+    """Implements the (basic) user interface for the Elliptical Slice sampling kernel.
+
+    Examples
+    --------
+
+    A new Elliptical Slice sampling kernel can be initialized and used with the following code:
+
+    .. code::
+
+        ellip_slice = blackjax.elliptical_slice(loglikelihood_fn, cov_matrix)
+        state = ellip_slice.init(position)
+        new_state, info = ellip_slice.step(rng_key, state)
+
+    We can JIT-compile the step function for better performance
+
+    .. code::
+
+        step = jax.jit(ellip_slice.step)
+        new_state, info = step(rng_key, state)
+
+    Parameters
+    ----------
+    loglikelihood_fn
+        Only the log likelihood function from the posterior distributon we wish to sample.
+    cov_matrix
+        The value of the covariance matrix of the gaussian prior distribution from the posterior we wish to sample.
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_kernel)
+
+    def __new__(  # type: ignore[misc]
+        cls,
+        loglikelihood_fn: Callable,
+        *,
+        mean: Array,
+        cov: Array,
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel(cov, mean)
+
+        def init_fn(position: PyTree):
+            return cls.init(position, loglikelihood_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(
+                rng_key,
+                state,
+                loglikelihood_fn,
+            )
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)
 
 
 def elliptical_proposal(
