+++ blackjax/mcmc/ghmc.py
@@ -21,10 +21,11 @@ import blackjax.mcmc.hmc as hmc
 import blackjax.mcmc.integrators as integrators
 import blackjax.mcmc.metrics as metrics
 import blackjax.mcmc.proposal as proposal
+from blackjax.base import MCMCSamplingAlgorithm
 from blackjax.types import PRNGKey, PyTree
 from blackjax.util import generate_gaussian_noise, pytree_size
 
-__all__ = ["GHMCState", "init", "build_kernel"]
+__all__ = ["GHMCState", "init", "build_kernel", "ghmc"]
 
 
 class GHMCState(NamedTuple):
@@ -167,6 +168,101 @@ def build_kernel(
         return state, info
 
     return kernel
+
+
+class ghmc:
+    """Implements the (basic) user interface for the Generalized HMC kernel.
+
+    The Generalized HMC kernel performs a similar procedure to the standard HMC
+    kernel with the difference of a persistent momentum variable and a non-reversible
+    Metropolis-Hastings step instead of the standard Metropolis-Hastings acceptance
+    step.
+
+    This means that the sampling of the momentum variable depends on the previous
+    momentum, the rate of persistence depends on the alpha parameter, and that the
+    Metropolis-Hastings accept/reject step is done through slice sampling with a
+    non-reversible slice variable also dependent on the previous slice, the determinisitc
+    transformation is defined by the delta parameter.
+
+    The Generalized HMC does not have a trajectory length parameter, it always performs
+    one iteration of the velocity verlet integrator with a given step size, making
+    the algorithm a good candiate for running many chains in parallel.
+
+    Examples
+    --------
+
+    A new Generalized HMC kernel can be initialized and used with the following code:
+
+    .. code::
+
+        ghmc_kernel = blackjax.ghmc(logdensity_fn, step_size, alpha, delta)
+        state = ghmc_kernel.init(rng_key, position)
+        new_state, info = ghmc_kernel.step(rng_key, state)
+
+    We can JIT-compile the step function for better performance
+
+    .. code::
+
+        step = jax.jit(ghmc_kernel.step)
+        new_state, info = step(rng_key, state)
+
+    Parameters
+    ----------
+    logdensity_fn
+        The log-density function we wish to draw samples from.
+    step_size
+        A PyTree of the same structure as the target PyTree (position) with the
+        values used for as a step size for each dimension of the target space in
+        the velocity verlet integrator.
+    alpha
+        The value defining the persistence of the momentum variable.
+    delta
+        The value defining the deterministic translation of the slice variable.
+    divergence_threshold
+        The absolute value of the difference in energy between two states above
+        which we say that the transition is divergent. The default value is
+        commonly found in other libraries, and yet is arbitrary.
+    noise_gn
+        A function that takes as input the slice variable and outputs a random
+        variable used as a noise correction of the persistent slice update.
+        The parameter defaults to a random variable with a single atom at 0.
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_kernel)
+
+    def __new__(  # type: ignore[misc]
+        cls,
+        logdensity_fn: Callable,
+        step_size: float,
+        momentum_inverse_scale: PyTree,
+        alpha: float,
+        delta: float,
+        *,
+        divergence_threshold: int = 1000,
+        noise_gn: Callable = lambda _: 0.0,
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel(noise_gn, divergence_threshold)
+
+        def init_fn(position: PyTree, rng_key: PRNGKey):
+            return cls.init(rng_key, position, logdensity_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(
+                rng_key,
+                state,
+                logdensity_fn,
+                step_size,
+                momentum_inverse_scale,
+                alpha,
+                delta,
+            )
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)  # type: ignore[arg-type]
 
 
 def update_momentum(rng_key, state, alpha):
