+++ blackjax/mcmc/mala.py
@@ -20,9 +20,10 @@ import jax.numpy as jnp
 
 import blackjax.mcmc.diffusions as diffusions
 import blackjax.mcmc.proposal as proposal
+from blackjax.base import MCMCSamplingAlgorithm
 from blackjax.types import PRNGKey, PyTree
 
-__all__ = ["MALAState", "MALAInfo", "init", "build_kernel"]
+__all__ = ["MALAState", "MALAInfo", "init", "build_kernel", "mala"]
 
 
 class MALAState(NamedTuple):
@@ -116,3 +117,72 @@ def build_kernel():
         return sampled_proposal.state, info
 
     return kernel
+
+
+class mala:
+    """Implements the (basic) user interface for the MALA kernel.
+
+    The general mala kernel builder (:meth:`blackjax.mcmc.mala.build_kernel`, alias `blackjax.mala.build_kernel`) can be
+    cumbersome to manipulate. Since most users only need to specify the kernel
+    parameters at initialization time, we provide a helper function that
+    specializes the general kernel.
+
+    We also add the general kernel and state generator as an attribute to this class so
+    users only need to pass `blackjax.mala` to SMC, adaptation, etc. algorithms.
+
+    Examples
+    --------
+
+    A new MALA kernel can be initialized and used with the following code:
+
+    .. code::
+
+        mala = blackjax.mala(logdensity_fn, step_size)
+        state = mala.init(position)
+        new_state, info = mala.step(rng_key, state)
+
+    Kernels are not jit-compiled by default so you will need to do it manually:
+
+    .. code::
+
+       step = jax.jit(mala.step)
+       new_state, info = step(rng_key, state)
+
+    Should you need to you can always use the base kernel directly:
+
+    .. code::
+
+       kernel = blackjax.mala.build_kernel(logdensity_fn)
+       state = blackjax.mala.init(position, logdensity_fn)
+       state, info = kernel(rng_key, state, logdensity_fn, step_size)
+
+    Parameters
+    ----------
+    logdensity_fn
+        The log-density function we wish to draw samples from.
+    step_size
+        The value to use for the step size in the symplectic integrator.
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_kernel)
+
+    def __new__(  # type: ignore[misc]
+        cls,
+        logdensity_fn: Callable,
+        step_size: float,
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel()
+
+        def init_fn(position: PyTree):
+            return cls.init(position, logdensity_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(rng_key, state, logdensity_fn, step_size)
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)
