+++ blackjax/mcmc/random_walk.py
@@ -59,6 +59,7 @@ import jax
 import numpy as np
 from jax import numpy as jnp
 
+from blackjax.base import MCMCSamplingAlgorithm
 from blackjax.mcmc import proposal
 from blackjax.types import Array, PRNGKey, PyTree
 from blackjax.util import generate_gaussian_noise
@@ -72,6 +73,9 @@ __all__ = [
     "RWState",
     "rmh_proposal",
     "build_rmh_transition_energy",
+    "additive_step_random_walk",
+    "irmh",
+    "rmh",
 ]
 
 
@@ -172,6 +176,75 @@ def build_additive_step():
     return kernel
 
 
+class additive_step_random_walk:
+    """Implements the user interface for the Additive Step RMH
+
+    Examples
+    --------
+
+    A new kernel can be initialized and used with the following code:
+
+    .. code::
+
+        rw = blackjax.additive_step_random_walk(logdensity_fn, random_step)
+        state = rw.init(position)
+        new_state, info = rw.step(rng_key, state)
+
+    The specific case of a Gaussian `random_step` is already implemented, either with independent components
+    when `covariance_matrix` is a one dimensional array or with dependent components if a two dimensional array:
+
+    .. code::
+
+        rw_gaussian = blackjax.additive_step_random_walk.normal_random_walk(logdensity_fn, covariance_matrix)
+        state = rw_gaussian.init(position)
+        new_state, info = rw_gaussian.step(rng_key, state)
+
+    Parameters
+    ----------
+    logdensity_fn
+        The log density probability density function from which we wish to sample.
+    random_step
+        A Callable that takes a random number generator and the current state and produces a step,
+        which will be added to the current position to obtain a new position. Must be symmetric
+        to maintain detailed balance. This means that P(step|position) = P(-step | position+step)
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_additive_step)
+
+    @classmethod
+    def normal_random_walk(cls, logdensity_fn: Callable, sigma):
+        """
+        Parameters
+        ----------
+        logdensity_fn
+            The log density probability density function from which we wish to sample.
+        sigma
+            The value of the covariance matrix of the gaussian proposal distribution.
+        Returns
+        -------
+             A ``MCMCSamplingAlgorithm``.
+        """
+        return cls(logdensity_fn, normal(sigma))
+
+    def __new__(  # type: ignore[misc]
+        cls, logdensity_fn: Callable, random_step: Callable
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel()
+
+        def init_fn(position: PyTree):
+            return cls.init(position, logdensity_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(rng_key, state, logdensity_fn, random_step)
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)
+
+
 def build_irmh() -> Callable:
     """
     Build an Independent Random Walk Rosenbluth-Metropolis-Hastings kernel. This implies
@@ -206,6 +279,60 @@ def build_irmh() -> Callable:
         return inner_kernel(rng_key, state, logdensity_fn, proposal_generator)
 
     return kernel
+
+
+class irmh:
+    """Implements the (basic) user interface for the independent RMH.
+
+    Examples
+    --------
+
+    A new kernel can be initialized and used with the following code:
+
+    .. code::
+
+        rmh = blackjax.irmh(logdensity_fn, proposal_distribution)
+        state = rmh.init(position)
+        new_state, info = rmh.step(rng_key, state)
+
+    We can JIT-compile the step function for better performance
+
+    .. code::
+
+        step = jax.jit(rmh.step)
+        new_state, info = step(rng_key, state)
+
+    Parameters
+    ----------
+    logdensity_fn
+        The log density probability density function from which we wish to sample.
+    proposal_distribution
+        A Callable that takes a random number generator and produces a new proposal. The
+        proposal is independent of the sampler's current state.
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_irmh)
+
+    def __new__(  # type: ignore[misc]
+        cls,
+        logdensity_fn: Callable,
+        proposal_distribution: Callable,
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel()
+
+        def init_fn(position: PyTree):
+            return cls.init(position, logdensity_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(rng_key, state, logdensity_fn, proposal_distribution)
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)
 
 
 def build_rmh():
@@ -264,6 +391,69 @@ def build_rmh():
         return new_state, RWInfo(p_accept, do_accept, new_state)
 
     return kernel
+
+
+class rmh:
+    """Implements the user interface for the RMH.
+
+    Examples
+    --------
+
+    A new kernel can be initialized and used with the following code:
+
+    .. code::
+
+        rmh = blackjax.rmh(logdensity_fn, proposal_generator)
+        state = rmh.init(position)
+        new_state, info = rmh.step(rng_key, state)
+
+    We can JIT-compile the step function for better performance
+
+    .. code::
+
+        step = jax.jit(rmh.step)
+        new_state, info = step(rng_key, state)
+
+    Parameters
+    ----------
+    logdensity_fn
+        The log density probability density function from which we wish to sample.
+    proposal_generator
+        A Callable that takes a random number generator and the current state and produces a new proposal.
+    proposal_logdensity_fn
+        The logdensity function associated to the proposal_generator. If the generator is non-symmetric,
+         P(x_t|x_t-1) is not equal to P(x_t-1|x_t), then this parameter must be not None in order to apply
+         the Metropolis-Hastings correction for detailed balance.
+
+    Returns
+    -------
+    A ``MCMCSamplingAlgorithm``.
+    """
+
+    init = staticmethod(init)
+    build_kernel = staticmethod(build_rmh)
+
+    def __new__(  # type: ignore[misc]
+        cls,
+        logdensity_fn: Callable,
+        proposal_generator: Callable[[PRNGKey, PyTree], PyTree],
+        proposal_logdensity_fn: Optional[Callable[[PyTree], PyTree]] = None,
+    ) -> MCMCSamplingAlgorithm:
+        kernel = cls.build_kernel()
+
+        def init_fn(position: PyTree):
+            return cls.init(position, logdensity_fn)
+
+        def step_fn(rng_key: PRNGKey, state):
+            return kernel(
+                rng_key,
+                state,
+                logdensity_fn,
+                proposal_generator,
+                proposal_logdensity_fn,
+            )
+
+        return MCMCSamplingAlgorithm(init_fn, step_fn)
 
 
 def build_rmh_transition_energy(proposal_logdensity_fn: Optional[Callable]) -> Callable:
